Moved code to mutation file
